+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 ## Unreleased
 
+### 4.3.18
+
+#### Dependency Upgrade
+- @twreporter/react-components@^7.1.1
+- @twreporter/react-article-components@1.0.20
+
+#### Rewrite components with scss with styled-components
+- Rewrite `AuthorsList`, `AuthorPage`, and `TopicLandingPage`
+- Rewrite `Photography`, and  `Search`
+
+#### Bug fix
+##### about-us page
+- Bugfix: member's carousel has wrong pagination on resize
+- Bugfix: side bar on about-us page
+
+#### Miscellaneous
+##### about-us page
+- `componentWillMount` -> `componentDidMount`, `componentWillUpdate` -> `shouldComponentUpdate`
+- Removes `shouldComponentUpdate` and `componentDidUpdate` in section1 since they are unused 
+  in current use case, which mounts a new component once animation get changed
+
 ## Release
 
 ### 4.3.17